requirements.txt
langchain
google-genai
fastapi
uvicorn
httpx
langchain-google-genai
langchain-core
langgraph
pydantic
python-dotenv
